t5505: remove sub shell use in favor of git -C
The t5505-remote.sh test script makes liberal use of sub shells to move
directories before executing a series of git commands. This is not the
typical style of a more modern test script. Instead, newer tests favor
the use of git's "-C" option to change directory before executing.

Update this test script to drop the unnecessary sub shells. Where
necessary, adjust the paths of various output files to account for this
change.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>