scalar: convert README.md into a technical design doc
Replace 'README.md' with 'technical/scalar.txt' (still in 'contrib/'). In
addition to reformatting for asciidoc, elaborate on the background, purpose,
and design choices that went into Scalar.

This document is intended to persist in the 'Documentation/technical/'
directory after Scalar has been moved into the root of Git (out of
'contrib/'). Before then, it will also contain a "Roadmap" section detailing
the remaining series needed to finish the initial version of Scalar. The
section will be removed once Scalar is moved to the repo root, but in the
meantime serves as a guide for readers to keep up with progress on the
feature.

Signed-off-by: Victoria Dye <vdye@github.com>