Documentation: remove use of whitelist
The word "whitelist" has cultural implications that are not inclusive.
Thankfully, it is not difficult to reword and avoid its use.

Remove uses of "whitelist" in the Documentation/ directory. In all
cases, we can rewrite the sentences to add clarity instead of relying on
the reader understanding this term.

The most substantial change is to git-daemon.txt which had several uses,
but we can refer to the input directories as a list of included
directories, making the descriptions slightly simpler.

Helped-by: Jeff King <peff@peff.net>
Helped-by: Junio C Hamano <gitster@pobox.com>
Signed-off-by: Derrick Stolee <derrickstolee@github.com>