rev-parse: respect push.autosetupremote when evaluating @{push}
In a previous release, the push.autosetupremote config was introduced to
ease new branch management in "simple" single-remote workflows. This makes
"git push" work on new branches (without configured upstream) with
push.default set to "simple" or "upstream" and it implies
"--set-upstream" regardless of whether the same-name remote branch exists
or not.

The "@{push}" suffix logic was not adjusted to account for this new option,
however, and sometimes returns an error when "git push" would successfully
push to an existing remote branch.

This is an edge-case, as the main context where push.autosetupremote will
apply is for *new* branches, with no corresponding remote branch yet, and
so even if the defaulting is handled correctly, the rev-parse will still
fail with "unknown revision or path not in the working tree".

Fix this edge-case so "git rev-parse @{push}" works, if there is no
upstream tracking relationship set up but the remote tracking branch that
will be defaulted to does already exist and can be resolved.

Also add corresponding test cases.

Signed-off-by: Tao Klerks <tao@klerks.biz>